requirements.txt
adjustText==0.7.3
aiofiles==0.7.0
aiosqlite==0.17.0
amqp==5.0.6
anyio==3.3.0
asgiref==3.4.1
atomicwrites==1.3.0
attrs==19.3.0
azure-core==1.17.0
azure-iot-hub==2.5.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
databases==0.4.3
defusedxml==0.7.1
Django==3.2.6
django-background-tasks==1.2.5
django-celery-results==2.2.0
django-compat==1.0.15
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flower==1.0.0
future==0.18.2
h11==0.12.0
humanize==3.11.0
idna==3.2
importlib-metadata==0.23
instance==0.0.0.0.0.0a0.dev20210120
iso8601==0.1.16
isodate==0.6.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.4.3
more-itertools==7.2.0
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
orm==0.2.0.dev1
packaging==19.2
pandas==1.3.2
patsy==0.5.1
Pillow==8.3.1
pluggy==0.13.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py==1.8.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.2
pytest==5.2.2
python-dateutil==2.8.2
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sprinkles==0.4.6
SQLAlchemy==1.3.24
sqlparse==0.4.1
starlette==0.16.0
statsmodels==0.12.2
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.62.1
typesystem==0.3.0.dev0
typing-extensions==3.10.0.0
uamqp==1.4.1
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zipp==0.6.0
